pyproject.toml
[tool.poetry]
name = "code-analysis-service"
version = "0.1.0"
description = "Service for analyzing code repositories"
authors = ["Parsa Harooni <prsishere@gmail.com.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
gitpython = "^3.1.44"
aiohttp = "^3.11.13"
sqlalchemy = "^2.0.28"
aiosqlite = "^0.20.0"
alembic = "^1.13.1"
greenlet = "^3.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.9.0"
pylint = "^3.1.0"
nox = "^2024.3.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pylint.messages_control]
disable = ["C0111", "C0103"]


